fix address gen test

+++ src/api/address_test.go
@@ -143,7 +143,7 @@ func TestGenerateAddresses(t *testing.T) {
 			var body GenerateAddressesRequest
 			err := json.Unmarshal([]byte(tc.httpBody), &body)
 			if err == nil {
-				gateway.On("AddressGen", body.AddressN, body.StartIndex, body.ConfirmAddress).Return(tc.gatewayAddressGenResult, nil)
+				gateway.On("AddressGen", uint32(body.AddressN), uint32(body.StartIndex), body.ConfirmAddress).Return(tc.gatewayAddressGenResult, nil)
 			}
 
 			req, err := http.NewRequest(tc.method, "/api/v1"+endpoint, strings.NewReader(tc.httpBody))
